pyproject.toml
[tool.poetry]
name = "powerk8s"
version = "0.1.0"
description = "Powerline Segment for Kubernetes"
authors = ["George Kontridze <george.kontridze@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
powerline-status = "^2.7"
kubernetes = "^12.0.1"
PyYAML = "^5.3.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
autoflake = "^1.4"
isort = "^5.6.4"
mypy = "^0.790"
ipython = "^7.19.0"
toml = "^0.10.2"
dephell = "^0.8.3"

[tool.poetry.scripts]
lint = "scripts.lint:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.autoflake]
recursive = true
in_place = true
ignore_init_module_imports = true
remove_all_unused_imports = true
remove_unused_variables = true
verbose = true
src = "."

[tool.dephell.main]
from = {format = "poetry", path = "pyproject.toml"}
to = {format = "setuppy", path = "setup.py"}
